requirements.txt
altair==4.2.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
Fiona==1.9.4.post1
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
isodate==0.6.1
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.24.3
oauth2client==4.1.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
plotly==5.15.0
protobuf==3.20.3
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
PyMySQL==1.0.3
pyparsing==3.0.9
pyproj==3.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
rich==13.3.5
rsa==4.9
shapely==2.0.1
six==1.16.0
smmap==5.0.0
streamlit==1.22.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.6.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.2
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
